# server/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ECHO API",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/v1/configuration": {
      "get": {
        "operationId": "getConfiguration",
        "tags": [
          "Public"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "minimumSupportedClientVersion": {
                      "type": "string"
                    },
                    "policyVersions": {
                      "additionalProperties": false,
                      "type": "object",
                      "properties": {
                        "privacy": {
                          "type": "string"
                        },
                        "terms": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "privacy",
                        "terms"
                      ]
                    },
                    "reflectionAvailable": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    },
                    "featureFlags": {
                      "additionalProperties": false,
                      "type": "object",
                      "properties": {
                        "backup": {
                          "type": "boolean"
                        },
                        "reflection": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "backup",
                        "reflection"
                      ]
                    }
                  },
                  "required": [
                    "minimumSupportedClientVersion",
                    "policyVersions",
                    "reflectionAvailable",
                    "featureFlags"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "format": "uuid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retryable",
                    "requestId"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/exchange": {
      "post": {
        "operationId": "exchangeProviderCredential",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "properties": {
                  "provider": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "apple"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "google"
                        ]
                      }
                    ]
                  },
                  "assertion": {
                    "minLength": 1,
                    "maxLength": 8192,
                    "type": "string"
                  }
                },
                "required": [
                  "provider",
                  "assertion"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "minimum": 1,
                      "type": "integer"
                    },
                    "tokenType": {
                      "type": "string",
                      "enum": [
                        "Bearer"
                      ]
                    },
                    "account": {
                      "additionalProperties": false,
                      "type": "object",
                      "properties": {
                        "id": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "displayName": {
                          "anyOf": [
                            {
                              "maxLength": 120,
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "locale": {
                          "type": "string"
                        },
                        "timezone": {
                          "type": "string"
                        },
                        "createdAt": {
                          "format": "date-time",
                          "type": "string"
                        },
                        "updatedAt": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "displayName",
                        "locale",
                        "timezone",
                        "createdAt",
                        "updatedAt"
                      ],
                      "title": "Account"
                    }
                  },
                  "required": [
                    "accessToken",
                    "refreshToken",
                    "expiresIn",
                    "tokenType",
                    "account"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "format": "uuid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retryable",
                    "requestId"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "format": "uuid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retryable",
                    "requestId"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "format": "uuid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retryable",
                    "requestId"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/magic-link/request": {
      "post": {
        "operationId": "requestMagicLink",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "properties": {
                  "email": {
                    "format": "email",
                    "maxLength": 320,
                    "type": "string"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "accepted": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  },
                  "required": [
                    "accepted"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "format": "uuid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retryable",
                    "requestId"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "format": "uuid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retryable",
                    "requestId"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/magic-link/verify": {
      "post": {
        "operationId": "verifyMagicLink",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "additionalProperties": false,
                    "type": "object",
                    "properties": {
                      "email": {
                        "format": "email",
                        "maxLength": 320,
                        "type": "string"
                      },
                      "token": {
                        "minLength": 1,
                        "maxLength": 8192,
                        "type": "string"
                      }
                    },
                    "required": [
                      "email",
                      "token"
                    ]
                  },
                  {
                    "additionalProperties": false,
                    "type": "object",
                    "properties": {
                      "tokenHash": {
                        "minLength": 1,
                        "maxLength": 8192,
                        "type": "string"
                      }
                    },
                    "required": [
                      "tokenHash"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "minimum": 1,
                      "type": "integer"
                    },
                    "tokenType": {
                      "type": "string",
                      "enum": [
                        "Bearer"
                      ]
                    },
                    "account": {
                      "additionalProperties": false,
                      "type": "object",
                      "properties": {
                        "id": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "displayName": {
                          "anyOf": [
                            {
                              "maxLength": 120,
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "locale": {
                          "type": "string"
                        },
                        "timezone": {
                          "type": "string"
                        },
                        "createdAt": {
                          "format": "date-time",
                          "type": "string"
                        },
                        "updatedAt": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "displayName",
                        "locale",
                        "timezone",
                        "createdAt",
                        "updatedAt"
                      ]
                    }
                  },
                  "required": [
                    "accessToken",
                    "refreshToken",
                    "expiresIn",
                    "tokenType",
                    "account"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "format": "uuid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retryable",
                    "requestId"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "format": "uuid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retryable",
                    "requestId"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "format": "uuid",
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "retryable",
                    "requestId"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}